python/requirements.txt
pandas
matplotlib
pyserial
coloredlogs
